WIP: Moving work to worker
