+++ dmci/api/api.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+class Api(Flask):
+    self.worker = None
