+++ tests/test_api/api_test_config.yaml
@@ -1,0 +1,6 @@
+api:
+  distributorQueuePaths: 
+    - temp
+  distribute:
+    - git
+    - pycsw
